fix: boardProvider and xi.hocus connect
