+++ packages/pkg.module.board/Board.tsx
@@ -17,6 +17,7 @@ type BoardPropsT = {
 };
 
 export const Board = ({ token }: BoardPropsT) => {
+  console.log('token', token);
   const store = useYjsStore({ roomId: token });
 
   return (
