--- packages/pkg.module.board/Provider.tsx
@@ -1,39 +1,0 @@
-'use client';
-
-import { useMainSt } from 'pkg.stores';
-import React from 'react';
-import './index.css';
-import { useParams } from 'next/navigation';
-import { toast } from 'sonner';
-import { Board } from './Board';
-
-export const Provider = () => {
-  const [token, setToken] = React.useState<null | string>(null);
-
-  const socket = useMainSt((state) => state.socket);
-  const params = useParams<{ 'channel-id': string; 'community-id': string }>();
-
-  React.useEffect(() => {
-    socket.emit(
-      'retrieve-board-channel',
-      {
-        community_id: params['community-id'],
-        channel_id: params['channel-id'],
-      },
-      (status: number, { ydoc_id: YdocId }: { ydoc_id: string }) => {
-        console.log('status', status);
-        if (status === 200) {
-          setToken(YdocId);
-        } else {
-          toast('Произошла ошибка при получении токена');
-        }
-      },
-    );
-  }, []);
-
-  if (token === null) {
-    return <span> Loading </span>;
-  }
-
-  return <Board token={token} />;
-};
