+++ packages/pkg.module.board/index.tsx
@@ -1,1 +1,2 @@
+export { BoardProvider } from './BoardProvider';
 export { Board } from './Board';
